+++ app/src/main/java/com/project/drinkly_admin/ui/MainActivity.kt
@@ -1,10 +1,11 @@
-package com.project.drinkly_admin
+package com.project.drinkly_admin.ui
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.project.drinkly_admin.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
