+++ app/src/main/java/com/project/drinkly_admin/ui/onboarding/SplashFragment.kt
@@ -1,0 +1,20 @@
+package com.project.drinkly_admin.ui.onboarding
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.drinkly_admin.R
+
+class SplashFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_splash, container, false)
+    }
+
+}
